+++ backend/gn_module_export/__init__.py
@@ -1,0 +1,2 @@
+MODULE_CODE = 'EXPORTS'
+MODULE_PICTO = 'fa-download'